requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.1
openai>=1.32.0
jinja2>=3.1.2
rich>=13.3.5
pydantic>=2.0.0
google-search-results>=2.4.2
playwright>=1.40.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
streamlit>=1.28.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pandas>=2.0.0
plotly>=5.17.0
